Syntax highlighting with highlightjs
The tricky parts:

- Highlighting after each update, since we're constantly adding/
removing pre/code nodes. Turned out not to be too tricky in the end.
- Language support. The latest published version of highlightjs,
9.12.0, does not have Reason highlighting support. So I forked it,
published the latest version to npm under a different name, and use
it in package.json.
- Light/dark mode support. There is no such thing as conditional
require in BS, so I load both the light and dark CSS files from
my highlightjs fork, and manage the two `style` DOM elements depending
on the app state. There must be a better way to do this :(